Moved models, ports and services in parent domain package
