+++ src/main/java/de/dhbw/karlsruhe/Main.java
@@ -1,26 +1,22 @@
 package de.dhbw.karlsruhe;
 
-import de.dhbw.karlsruhe.models.Difficulty;
-import de.dhbw.karlsruhe.models.Sudoku;
-import de.dhbw.karlsruhe.services.DifficultySelectionDialogService;
-import de.dhbw.karlsruhe.services.StartUpDialogService;
-import de.dhbw.karlsruhe.services.LogoutService;
-import de.dhbw.karlsruhe.services.MenuDialogService;
+import de.dhbw.karlsruhe.domain.services.LogoutService;
+import de.dhbw.karlsruhe.domain.services.MenuDialogService;
+import de.dhbw.karlsruhe.domain.services.StartUpDialogService;
 
 public class Main {
 
-	public static void main(String[] args) {
-		boolean desireToRun = true;
-		LogoutService logoutService = new LogoutService();
-		StartUpDialogService startUpDialogService = new StartUpDialogService();
-        MenuDialogService menuDialogService = new MenuDialogService();
+  public static void main(String[] args) {
+    boolean desireToRun = true;
+    LogoutService logoutService = new LogoutService();
+    StartUpDialogService startUpDialogService = new StartUpDialogService();
+    MenuDialogService menuDialogService = new MenuDialogService();
 
+    do {
+      startUpDialogService.signIn();
+      menuDialogService.startMenuDialog();
 
-		do {
-			startUpDialogService.signIn();
-			menuDialogService.startMenuDialog();
-
-			desireToRun = logoutService.checkDesireToRun();
-		} while (desireToRun);
-	}
+      desireToRun = logoutService.checkDesireToRun();
+    } while (desireToRun);
+  }
 }
