+++ src/main/java/de/dhbw/karlsruhe/domain/services/EncryptionService.java
@@ -1,4 +1,4 @@
-package de.dhbw.karlsruhe.services;
+package de.dhbw.karlsruhe.domain.services;
 
 import java.math.BigInteger;
 import java.nio.charset.StandardCharsets;
@@ -16,8 +16,7 @@ public class EncryptionService {
     BigInteger number = new BigInteger(1, hashValue);
     StringBuilder hexString = new StringBuilder(number.toString(16));
 
-    while (hexString.length() < 32)
-    {
+    while (hexString.length() < 32) {
       hexString.insert(0, '0');
     }
