+++ src/main/java/de/dhbw/karlsruhe/domain/services/LeaderboardDialogService.java
@@ -1,0 +1,12 @@
+package de.dhbw.karlsruhe.domain.services;
+
+public class LeaderboardDialogService {
+
+  public LeaderboardDialogService() {
+
+  }
+
+  public void startLeaderboardDialog() {
+    System.out.println("Leaderboard Dialog");
+  }
+}
