--- src/main/java/de/dhbw/karlsruhe/ports/SudokuPort.java
@@ -1,16 +1,0 @@
-package de.dhbw.karlsruhe.ports;
-
-import de.dhbw.karlsruhe.models.Sudoku;
-
-import java.util.List;
-
-public interface SudokuPort {
-
-    void saveSudoku(Sudoku sudoku);
-
-    List<Sudoku> getAllSudokus();
-
-    Sudoku getSudoku(String id);
-
-    void deleteSudoku(String id);
-}
