--- src/main/java/de/dhbw/karlsruhe/services/LeaderboardDialogService.java
@@ -1,12 +1,0 @@
-package de.dhbw.karlsruhe.services;
-
-public class LeaderboardDialogService {
-
-    public LeaderboardDialogService() {
-
-    }
-
-    public void startLeaderboardDialog() {
-        System.out.println("Leaderboard Dialog");
-    }
-}
