--- src/main/java/de/dhbw/karlsruhe/services/ScannerService.java
@@ -1,15 +1,0 @@
-package de.dhbw.karlsruhe.services;
-
-import java.util.Scanner;
-
-public class ScannerService {
-    private static Scanner scanner ;
-
-    public static Scanner getScanner() {
-        if (scanner == null) {
-            scanner = new Scanner(System.in);
-        }
-
-        return scanner;
-    }
-}
